[STB] Implementing the UserEvent mechanism for receiving RCU keys (MUST: See details)
Details:
This is done since there was an issue reported in Irdeto middleware that the
events are not delivered to the xlets htrough the AWT mechanism even though
the xlet added the keyListener and requested the focus on the widget.
Note that if the AWT mechanism is working, there will be two events obtained
to the xlet with the same key code (one thru UserEvent and other thru AWT key
event), which will cause erraneous behavior. In such cases, either one method
needs to be commented out.